Added a Camera to the game. We can move it with the arrows from the keyboard. We can also zoom in or zoom out with Z and S

+++ interface/grid.h
@@ -16,7 +16,7 @@ private:
 public:
     Grid(Hexagrid hexagrid);
     std::shared_ptr<Hexagon> getHexagon(std::shared_ptr<Cell> cell);
-    sf::Vector2f PixToCell(int xCursor, int yCursor);
+    sf::Vector2f PixToCell(float xCursor, float yCursor);
     void update(std::string tide);
     void draw(sf::RenderTarget& target, sf::RenderStates states) const;
 };
